requirements.txt
anyio==3.6.2
appnope==0.1.3
attrs==23.1.0
backcall==0.2.0
bleach==6.0.0
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
click-log==0.4.0
colorama==0.4.6
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
docutils==0.20
dotty-dict==1.3.1
entrypoints==0.4
fastapi==0.95.1
gitdb==4.0.10
GitPython==3.1.31
gotrue==0.5.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.16.3
httpx==0.23.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
invoke==1.7.3
ipykernel==6.16.2
ipython==7.34.0
jaraco.classes==3.2.3
jedi==0.18.2
joblib==1.3.1
jupyter_client==7.4.9
jupyter_core==4.12.0
keyring==23.13.1
lxml==4.9.3
matplotlib-inline==0.1.6
more-itertools==9.1.0
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.21.6
packaging==23.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.9.6
pluggy==1.0.0
postgrest-py==0.10.4
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
py==1.11.0
pydantic==1.10.7
Pygments==2.15.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.21.1
python-gitlab==3.14.0
python-semantic-release==7.32.1
pytrends @ git+https://github.com/GeneralMills/pytrends@eb84a3b745359029d784334e735bc1a0e7a77e83
pytz==2023.3
pyzmq==25.0.2
readme-renderer==37.3
realtime==0.0.5
realtime-py==0.1.3
regex==2023.6.3
requests==2.25.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.26.1
storage3==0.3.6
StrEnum==0.4.10
supabase==0.7.1
supafunc==0.2.2
textblob==0.17.1
toml==0.10.2
tomlkit==0.11.8
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
twine==3.8.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.22.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
